update code to remove warnings (mostly type conflicts) in visual studio

+++ Brain/MarkovBrain/GateBuilder/GateBuilder.cpp
@@ -154,7 +154,7 @@ void Gate_Builder::setupGates() {
 			string IO_Ranges = (_PT == nullptr) ? ProbabilisticGate::IO_RangesPL->lookup() : _PT->lookupString("BRAIN_MARKOV_GATES_PROBABILISTIC-IO_Ranges");
 			int maxIn, maxOut;
 			pair<vector<int>,vector<int>> addresses = getInputsAndOutputs(IO_Ranges, maxIn, maxOut, genomeHandler, gateID,"BRAIN_MARKOV_GATES_PROBABILISTIC");
-			vector<vector<int>> rawTable = genomeHandler->readTable( {1 << addresses.first.size(), 1 << addresses.second.size()}, {pow(2,maxIn), pow(2,maxOut)}, {0, 255}, AbstractGate::DATA_CODE, gateID);
+			vector<vector<int>> rawTable = genomeHandler->readTable( {1 << addresses.first.size(), 1 << addresses.second.size()}, {(int)pow(2,maxIn), (int)pow(2,maxOut)}, {0, 255}, AbstractGate::DATA_CODE, gateID);
 			if (genomeHandler->atEOC()) {
 				shared_ptr<ProbabilisticGate> nullObj = nullptr;
 				return nullObj;
@@ -176,7 +176,7 @@ void Gate_Builder::setupGates() {
 			string IO_Ranges = (_PT == nullptr) ? DeterministicGate::IO_RangesPL->lookup() : _PT->lookupString("BRAIN_MARKOV_GATES_DETERMINISTIC-IO_Ranges");
 			int maxIn, maxOut;
 			pair<vector<int>,vector<int>> addresses = getInputsAndOutputs(IO_Ranges, maxIn, maxOut, genomeHandler, gateID,"BRAIN_MARKOV_GATES_DETERMINISTIC");
-			vector<vector<int>> table = genomeHandler->readTable( {1 << addresses.first.size(), addresses.second.size()}, {pow(2,maxIn), maxOut}, {0, 1}, AbstractGate::DATA_CODE, gateID);
+			vector<vector<int>> table = genomeHandler->readTable( {1 << (int)addresses.first.size(), (int)addresses.second.size()}, {(int)pow(2,maxIn), maxOut}, {0, 1}, AbstractGate::DATA_CODE, gateID);
 			if (genomeHandler->atEOC()) {
 				shared_ptr<DeterministicGate> nullObj = nullptr;
 				return nullObj;
@@ -200,7 +200,7 @@ void Gate_Builder::setupGates() {
 			string IO_Ranges = (_PT == nullptr) ? EpsilonGate::IO_RangesPL->lookup() : _PT->lookupString("BRAIN_MARKOV_GATES_EPSILON-IO_Ranges");
 			int maxIn, maxOut;
 			pair<vector<int>,vector<int>> addresses = getInputsAndOutputs(IO_Ranges, maxIn, maxOut, genomeHandler, gateID,"BRAIN_MARKOV_GATES_EPSILON");
-			vector<vector<int>> table = genomeHandler->readTable( {1 << addresses.first.size(), addresses.second.size()}, {pow(2,maxIn), maxOut}, {0, 1}, AbstractGate::DATA_CODE, gateID);
+			vector<vector<int>> table = genomeHandler->readTable( {1 << (int)addresses.first.size(), (int)addresses.second.size()}, {(int)pow(2,maxIn), maxOut}, {0, 1}, AbstractGate::DATA_CODE, gateID);
 
 			double epsilon = (_PT == nullptr) ? EpsilonGate::EpsilonSourcePL->lookup() : _PT->lookupDouble("BRAIN_MARKOV_GATES_EPSILON-epsilonSource");
 
@@ -235,7 +235,7 @@ void Gate_Builder::setupGates() {
 			string IO_Ranges = (_PT == nullptr) ? VoidGate::IO_RangesPL->lookup() : _PT->lookupString("BRAIN_MARKOV_GATES_VOID-IO_Ranges");
 			int maxIn, maxOut;
 			pair<vector<int>,vector<int>> addresses = getInputsAndOutputs(IO_Ranges, maxIn, maxOut, genomeHandler, gateID,"BRAIN_MARKOV_GATES_VOID");
-			vector<vector<int>> table = genomeHandler->readTable( {1 << addresses.first.size(), addresses.second.size()}, {pow(2,maxIn), maxOut}, {0, 1}, AbstractGate::DATA_CODE, gateID);
+			vector<vector<int>> table = genomeHandler->readTable( {1 << (int)addresses.first.size(), (int)addresses.second.size()}, {(int)pow(2,maxIn), maxOut}, {0, 1}, AbstractGate::DATA_CODE, gateID);
 
 			double epsilon = (_PT == nullptr) ? VoidGate::voidGate_ProbabilityPL->lookup() : _PT->lookupDouble("BRAIN_MARKOV_GATES_VOID-epsilonSource");
 
@@ -335,7 +335,7 @@ void Gate_Builder::setupGates() {
 			string IO_Ranges = (_PT == nullptr) ? TritDeterministicGate::IO_RangesPL->lookup() : _PT->lookupString("BRAIN_MARKOV_GATES_TRIT-IO_Ranges");
 			int maxIn, maxOut;
 			pair<vector<int>,vector<int>> addresses = getInputsAndOutputs(IO_Ranges, maxIn, maxOut, genomeHandler, gateID,"BRAIN_MARKOV_GATES_TRIT");
-			vector<vector<int>> table = genomeHandler->readTable( {pow(3,addresses.first.size()), addresses.second.size()}, {pow(3,maxIn), maxOut}, {-1, 1}, AbstractGate::DATA_CODE, gateID);
+			vector<vector<int>> table = genomeHandler->readTable( {(int)pow(3,(int)addresses.first.size()), (int)addresses.second.size()}, {(int)pow(3,maxIn), maxOut}, {-1, 1}, AbstractGate::DATA_CODE, gateID);
 			if (genomeHandler->atEOC()) {
 				shared_ptr<TritDeterministicGate> nullObj = nullptr;;
 				return nullObj;
